Limit number of requests waiting at the gate
In order to avoid a high numbers of in-flight requests in case requests
are handled at a slower rate than they arrive, add a limit to how many
requests are allowed to wait at the gate. If the wait limit has been
reached and a new request arrives, gate will respond HTTP 429 directly.